Put utility functions in one file.
