+++ src/components/custom/ProfilesTable.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { formatDate } from "../../util/dateFormatter";
+import { buildFullName, formatDate } from "../../utils/appUtils";
 
 import { PROFILE_PATH } from "../../constants/app.constants";
 
@@ -29,7 +29,13 @@ const ProfilesTable: React.FC<ProfileTableProps> = ({ profiles }) => {
                   </Link>
                 }
               </td>
-              <td>{`${profile.lastName}, ${profile.middleName} ${profile.firstName}`}</td>
+              <td>
+                {buildFullName(
+                  profile.firstName,
+                  profile.middleName,
+                  profile.lastName,
+                )}
+              </td>
               <td>{formatDate(profile.createdDateTime)}</td>
             </tr>
           ))}
