--- src/util/dateFormatter.ts
@@ -1,5 +1,0 @@
-const formatDate = (date: string): string => {
-  return date.split(".")[0].replace("T", " ");
-};
-
-export { formatDate };
